Improved shut-down/state management
We notice that if the acpid process does not start this prevents
hyperkit from exiting cleanly when it is sent a SIGTERM. In order to
make the tests more reliable we now capture the linuxkit pid and kill
the entire process group. The stop command does this as well.

Given that this results in a dirty state directory the start command
will now clear this directory when starting a new vm instance

tl;dr
- the state directory is always deleted
- we write the linuxkit pid and stop command uses it to kill the process group

Signed-off-by: Steve Hiehn <shiehn@pivotal.io>