Deploy CF
We created a CF iso with the necessary dependencies. The CLI has been
updated to drive the deploy via a garden container.

Since garden containers do not have access to the host networking we
created a linuxkit package (setup-bosh-lite-routing) to poll for when
the director is up. In order to make the nested containers (inside the
BOSH director container) routable.

Signed-off-by: Steve Hiehn <shiehn@pivotal.io>