Fix network stability issues when BOSH is not deployed
bosh-lite-routing would loop and curl the Director IP which was actually
routed outside of the VM. Due to the loopback alias on the Mac, it would be
routed back into the VM. vpnkit-forwarder in the VM would then make the
exact same request causing us to create alot of TCP connections. These
connections would remain in a ESTABLISHED/CLOSE_WAIT for quite some
time.

We now signal when the director is up by touching a file in a shared
director. Hence we should no longer see EOFs occuring from the cfdev CLI
when attempting to start deploying the BOSH director & CF.

[#153674145]

Signed-off-by: Steve Hiehn <shiehn@pivotal.io>