Add Acceptance test for telemetry turn off
For the implementation of this test we changed the function that
retrieves the information from Kinesis.
Now we store in a global state all the metrics that we need to check
instead of using a single boolean.
This changed the signature of the function.

Now we can check any metric that we want.